requirements.txt
google-adk
python-dotenv
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6